+++ product-catalog/test/productcatalog/pages/ProductDataFactoryTest.java
@@ -44,36 +44,6 @@ public class ProductDataFactoryTest {
     }
 
     private ReverseRouter reverseRouter() {
-        return new ReverseRouter() {
-            @Override
-            public Call home(final String language) {
-                return null;
-            }
-
-            @Override
-            public Call category(final String locale, final String slug, final int page) {
-                return null;
-            }
-
-            @Override
-            public Call product(final String locale, final String productSlug, final String sku) {
-                return new Call() {
-                    @Override
-                    public String url() {
-                        return productSlug + "-url-" + sku;
-                    }
-
-                    @Override
-                    public String method() {
-                        return null;
-                    }
-
-                    @Override
-                    public String fragment() {
-                        return null;
-                    }
-                };
-            }
-        };
+        return new DefaultTestReverseRouter();
     }
 }
